+++ tests/simple/test_simple.py
@@ -22,6 +22,8 @@ import typing
 
 import pytest
 
+from tests.conftest import CONDITION_COVERAGE_POSSIBLE
+
 
 if typing.TYPE_CHECKING:
     from tests.conftest import GcovrTestExec
@@ -308,15 +310,25 @@ def test_fail_under(
 
     process = gcovr_test_exec.gcovr(
         "--decision",
-        "--fail-under-decision=50.1",
+        "--fail-under-condition-or-decision=50.1",
         "--print-summary",
         use_main=True,
     )
     assert process.returncode == 8
+    # The condition coverage is only reported if the compiler supports it.
+    expected_messages = [
+        *(
+            ["Failed minimum condition coverage "]
+            if CONDITION_COVERAGE_POSSIBLE
+            else []
+        ),
+        "Failed minimum decision coverage ",
+    ]
     messages = caplog.record_tuples
-    assert len(messages) == 1
-    assert messages[0][1] == logging.ERROR
-    assert messages[0][2].startswith("Failed minimum decision coverage ")
+    assert len(messages) == len(expected_messages)
+    for message, expected_message in zip(messages, expected_messages, strict=True):
+        assert message[1] == logging.ERROR
+        assert message[2].startswith(expected_message)
     caplog.clear()
 
     process = gcovr_test_exec.gcovr(
@@ -335,22 +347,28 @@ def test_fail_under(
         "--fail-under-line=80.1",
         "--fail-under-branch=50.1",
         "--decision",
-        "--fail-under-decision=50.1",
+        "--fail-under-condition-or-decision=50.1",
         "--fail-under-function=66.8",
         "--print-summary",
         use_main=True,
     )
     assert process.returncode == 30
+    expected_messages = [
+        "Failed minimum line coverage ",
+        "Failed minimum branch coverage ",
+        *(
+            ["Failed minimum condition coverage "]
+            if CONDITION_COVERAGE_POSSIBLE
+            else []
+        ),
+        "Failed minimum decision coverage ",
+        "Failed minimum function coverage ",
+    ]
     messages = caplog.record_tuples
-    assert len(messages) == 4
-    assert messages[0][1] == logging.ERROR
-    assert messages[0][2].startswith("Failed minimum line coverage ")
-    assert messages[1][1] == logging.ERROR
-    assert messages[1][2].startswith("Failed minimum branch coverage ")
-    assert messages[2][1] == logging.ERROR
-    assert messages[2][2].startswith("Failed minimum decision coverage ")
-    assert messages[3][1] == logging.ERROR
-    assert messages[3][2].startswith("Failed minimum function coverage ")
+    assert len(messages) == len(expected_messages)
+    for message, expected_message in zip(messages, expected_messages, strict=True):
+        assert message[1] == logging.ERROR
+        assert message[2].startswith(expected_message)
     caplog.clear()
 
     process = gcovr_test_exec.gcovr(
@@ -358,7 +376,7 @@ def test_fail_under(
         "61.5" if gcovr_test_exec.is_llvm() else "63.6",
         "--fail-under-branch=50.0",
         "--decision",
-        "--fail-under-decision=50.0",
+        "--fail-under-condition-or-decision=50.0",
         "--fail-under-function=66.7",
         "--print-summary",
         use_main=True,
@@ -366,5 +384,6 @@ def test_fail_under(
     assert process.returncode == 0
     assert "(ERROR) Failed minimum line coverage" not in process.stderr
     assert "(ERROR) Failed minimum branch coverage" not in process.stderr
+    assert "(ERROR) Failed minimum condition coverage" not in process.stderr
     assert "(ERROR) Failed minimum decision coverage" not in process.stderr
     assert "(ERROR) Failed minimum function coverage" not in process.stderr